requirements.txt
# Core Requirements
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
keras==3.8.0
opencv-python-headless==4.8.0.76
flask==3.0.0
numpy==1.24.4
pillow==10.0.1
scikit-learn==1.3.2
joblib==1.3.2

# Mac-specific
pyobjc==9.2
pyobjc-framework-AVFoundation==9.2

# Training Compatibility
h5py==3.13.0
protobuf==3.20.3